Now, when "make" fails to certify a book, it prints:
   1. The **CERTIFICATION FAILED** line, as usual; 
   2. The last 300 lines of the .out file, nicely indented; and
   3. The **CERTIFICATION FAILED** line, again, so you can see which 
      file had the problem without scrolling back up.

To try this out, just add an "i-am-here" to any random book and try to certify
it.

I think this is convenient in the course of regular book development, because
it often lets me see why a book is failing without needing to open the .out
file or manually load the book.

It may also be very useful in automated regression scripts which send you an
email when they fail.  That is, often the last 300 lines are sufficient to see
why a book is failing, so as the recipient of such an email may not need to log
into a remote machine and look at a .out file.



git-svn-id: http://acl2-books.googlecode.com/svn/trunk@188 303c5b2c-9740-0410-b014-fd8f9db7be32