The main changes are summarized in the following lines taken from (deflabel note-4-2 ...).
.......

; A potential soundness hole was plugged in the proof-checker by making
; variable pc-assign untouchable.  But we don't mention this in the release
; notes proper because we have not been able to exploit this potential hole.

; Changed fmt-abbrev1 to print the message about :DOC set-iprint in column 0
; after a newline, because otherwise the new message printed immediately after
; evaluating (retrieve ...) looks odd.

Improved proof output in the case of a ~c[:]~ilc[clause-processor] hint that
proves the goal, so that the clause-processor function name is printed.

The ~ilc[proof-checker] command `~c[then]' now stops at the first failure (if
any).
 
A new hint, ~c[:]~ilc[instructions], allows use of the ~il[proof-checker] at
the level of ~il[hints] to the prover.  Thanks to Pete Manolios for
requesting this feature (in 2001!).  ~l[instructions].

Fixed a long-standing soundness bug caused by the interaction of ~ilc[force]d
hypotheses with destructor ~il[elim]ination.  The fix was to avoid using
forcing when building the context (so-called ``type-alist'') when the goal is
considered for destructor elimination; those who are interested can see a
discussion in source function ~c[eliminate-destructors-clause1], which
includes a proof of ~c[nil] that no longer succeeds.  A similar fix was made
for generalization, though we have not exploited the previous code to prove
~c[nil] in that case.


git-svn-id: http://acl2-devel.googlecode.com/svn/acl2svn/trunk@106 5d84a466-a78b-0930-a311-59b22171a541