(HONS version) Gag-mode now is initially set to :goals instead of t.
Removed support for "tainted" certificates.  One reason is that there
are rarely incremental releases.  A stronger reason is that for the
compatibility of a new release is with the previous non-incremental
release, it's not particularly relevant whether or not the new release
is incremental.

Made trivial wording change ("compiling" to "compile") in "Step 5"
certification message.


git-svn-id: http://acl2-devel.googlecode.com/svn/acl2svn/trunk@498 5d84a466-a78b-0930-a311-59b22171a541