Proposed changes to countereg-gen for Makefile-generic and cert.pl compatibility.
I'm not sure if this is going to work with ACL2s, so please be careful when updating
if you are using ACL2s.

Summary of changes:

- Added typical portcullis.acl2, portcullis.lisp, cert.acl2 files

- Changed pkg.lsp to package.lsp to match the convention of other libraries

- Added acl2-customization.lsp file to load package.lsp for convenience

- Replaced Makefile with Makefile-generic based replacement

- Changed ACL2s preambles in many files to just LD "cert.acl2", using a
  linebreak to fool Makefile-generic/cert.pl scanners

- Removed now unnecessary acl2s-lib.jar file

Note to Harsh: In random.lisp I had to make some updates to
keyword-package-intern, a lemma used to prove random-keyword-keyword,
apparently because of differences to the axiom KEYWORD-PACKAGE.  If this breaks
your copy, we will have to figure out what's happened between our ACL2
versions.  It looks like in acl2-sources/trunk the axiom has no variables, so
it may be that you just need to "svn up" your ACL2 to match.  (I hope).




git-svn-id: http://acl2-books.googlecode.com/svn/trunk@1111 303c5b2c-9740-0410-b014-fd8f9db7be32