Quoting :doc note-4-4:
| We now support ACL2(h) built not only on 64-bit CCL but also on 64-bit SBCL
| and 64-bit CMUCL.  Thanks to Jared Davis for doing most of the necessary
| work on this.  Other host Lisps may work as well; users who want such support
| but find it lacking are welcome to ask for it.

There are a few related minor mods.  For example, acl2-pc::nil is
excluded from memoization by dubious-to-profile, and
function-lambda-expression is avoided for functions that are memoized
by default.  Also, thanks to Jared Davis, eliminated a lot of trailing
whitespace in memoize-raw.lisp.

Also somewhat related to the above, improved memoization for
worse-than -- actually, worse-than-builtin -- so that its memo-table
is cleared just before and just after each proof.  I think this was
necessary in order to get the ACL2(h) regression to pass using CMUCL
as the host Lisp.  (Actually I'll probably remove an unnecessary call
in prove at a future time, since this is already covered in
prove-loop.)

Quoting :doc note-4-4:

| (SBCL only) ACL2 images built on SBCL now have an option,
| --dynamic-space-size 2000, that can avoid space problems that could
| previously have caused the session to die.


git-svn-id: http://acl2-devel.googlecode.com/svn/acl2svn/trunk@724 5d84a466-a78b-0930-a311-59b22171a541