Fixes for building ACL2(h) under Lispworks, some of which may apply to other ANSI Lisps other than CCL:
(1) I've put #+Clozure above the definition of watch-dump, because it
calls with-timer-raw-condition-type, which is only defined under
#+Clozure.  FYI, I discovered it when building ACL2(h) under
Lispworks:

;;;*** Warning in ACL2::WATCH-DUMP: ACL2::WITH-TIMER-RAW-CONDITION-TYPE is not a valid type.

(2) I got this warning (several times) in that Lispworks build, and to
fix it, I redefined GLOBLET to be LET when #-Clozure.

;;;*** Warning in ACL2::HONS-INIT-HOOK-MEMOIZATIONS: #:*COUNT-PONS-CALLS* bound lexically.

(3) Made a bizarre adjustment in with-lower-overhead (and to be
analogous, in lower-overhead), where it turns out that binding
*record-calls* to t instead of nil makes a huge positive difference.
See the comment there for an example.  Some day it would be good to
figure this out, but for now, given that the solution provides
reasonable behavior (and is only relevant to LispWorks builds of
ACL2(h)) and we are trying to complete the release, this is good
enough.


git-svn-id: http://acl2-devel.googlecode.com/svn/acl2svn/trunk@735 5d84a466-a78b-0930-a311-59b22171a541