Reorganizing many books.
(1) Added new directory "std" which will serve as the future container
    for all libraries that are deemed to be part of a standard way of
    enabling ACL2 to do powerful things beyond what it does just out
    of the box.  On a side note, this is mostly David and Jared's
    vision.

(2) Moved many libraries out of unicode (and into std/io, std/ks (ks
    is short for kitchen-sink), and std/lists).  This is long overdue,
    but we've known that it would break the build for many people, so
    we've been putting it off.  To Rager's knowledge, the books
    repository builds to completion using GNU Make and cert.pl to
    clean up the files that aren't part of the GNU Make system.

Sorry about potentially breaking the builds of any company that uses
bleeding-edge books for everyone overnight, but since 6.0 has been
released recently, this seems like a good time to do it.  Our hope is
that this is the start of a cleanup that will clearly benefit the ACL2
community in the medium and long-term.

git-svn-id: http://acl2-books.googlecode.com/svn/trunk@1436 303c5b2c-9740-0410-b014-fd8f9db7be32