In GNUmakefile, removed support for regressions that use infix printing.  The reason is that there was a `make' variable, ACL2_VERSION, that had been set to v5-0 (!).  That variable was only used by the infix-related targets, so I thought it best just to remove the whole thing, since I'm far from sure that we support infix printing now but even if we do, we never run regressions using it.
Fixed a bug for Lisps other than CCL and SBCL, in handling of mfc-xx
functions (such as mfc-ts-fn).  This bug was introduced recently when
implementing these using trusted clause processors.  The problem was
that comp-fn is called near the end of the build process when
(f-get-global 'compiler-enabled state) is not T (note that it is
:BOOKS for CCL and SBCL), and comp-fn was erroneously overwriting the
definitions of the *1* functions for the mfc-ts functions.  These had
been modified by the calls of the defun-overrides macro, but those
modifications were being lost.  The solution was to exempt
*defun-overrides* from compilation by compile-uncompiled-*1*-defuns,
much as they were already exempted in add-trip (though I've made a
minor aesthetic syntax improvement there).

Made small comment updates, in particular clarifying that CCL is the
Lisp of choice for ACL2(h).


git-svn-id: http://acl2-devel.googlecode.com/svn/acl2svn/trunk@908 5d84a466-a78b-0930-a311-59b22171a541