Just some small clean-up this time:
-----

Eliminated comments about function elide-locals-post, which no longer
exists.

Restricted special handling for gcl in GNUmakefile to the non-ANSI case.

Replaced :doc apropos by :doc finding-documentation, which is more
clearly referenced when :doc fails.  This new topic discusses the xdoc
manual.

I think that a recent change to the ACL2 build process was not
reflected in quick-compile-acl2, so I added the following comment to
(deflabel note-6-3 ...):

; Removed support for "make fast", "make very-fast", and "make fast-meter",
; which as far as we know are no longer in use.  Thus, deleted
; quick-compile-acl2.  Also deleted no-object-file-or-out-of-date-object-file,
; since it was there only to support quick-compile-acl2.


git-svn-id: http://acl2-devel.googlecode.com/svn/trunk@1083 5d84a466-a78b-0930-a311-59b22171a541