This is probably the final commit before inviting the ACL2 community to contribute to the ACL2 documentation.
Pending work includes:

- Modifications to the generated file manual/LICENSE that Jared
  intends to make, which clarify that it only applies to the viewer,
  not the actual doc strings.

- Modifications to the broken link mechanism, so that for example
  clicking on [xdoc] in :doc note-6-4 will take you to a nice page
  that suggests possibly using the combined acl2+books manual
  (see :doc broken-link).

In general, there are many opportunities for improving the ACL2 system
documentation.  So far, it's only slightly modified from a direct
translation of the original ACL2 documentation.  But for example, I
modified the form (defxdoc hints ...) in
books/system/doc/acl2-doc.lisp so to use <dl> ... </dl>, which makes
for a nicer display, and to fix (manually) the mixed lower/upper case
of the hint keywords listed therein.


git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2207 303c5b2c-9740-0410-b014-fd8f9db7be32