Updated Makefile in consultation with Jared, so that "clean" does "milawa-clean".
Also with Jared's help, updated top topic of acl2+books combined
manual so that the date is included.  I wanted because I now
distribute an svn version of the acl2+books combined manual from time
to time (see
http://www.cs.utexas.edu/users/moore/acl2/manuals/current/manual/),
and it might be nice for people to see the date on the manual that
they are viewing.


git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2248 303c5b2c-9740-0410-b014-fd8f9db7be32