Fixed :doc acl2-doc to reflect the following changes, about to be checked into acl2-devel:
Improved the ACL2-Doc "i" (index) command so that if you just hit
<RETURN>, then you go to the index buffer, which now is read-only and
is in ACL2-Doc mode (so that, for example, you can hit <RETURN> while
standing on a topic in order to be taken to that topic).

Modified :DOC broken-link for the ACL2-only manual so that it gives
specific information about the source books for missing documentation
links.


git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2375 303c5b2c-9740-0410-b014-fd8f9db7be32