Updated a bit of documentation relevant to sys-call and sys-call+, and added to :doc note-6-5 to reflect ACL2 changes:
  The utility [sys-call+] can now only be called if there is an active
  [trust-tag], which is the restriction that was already in place for
  [sys-call]. This plugs a potential soundness hole.

  The default value for the :write-port keyword argument of
  [certify-book] is now t, as was already claimed by the
  documentation. Thus, a .port file will be written when a book is
  certified, making it more likely that a subsequent [include-book]
  will complete without error even if the book has been modified
  without recertification. Note that a change by Sol Swords to the
  build system for the [community-books] now sets an environment
  variable to guarantee that write-port is true, independent of this
  change.  On a related note: the unadvertised environment
  variable ACL2_WRITE_PORT is now handled more appropriately, by
  interning its upcased argument into the ACL2 package.

Because of the above sys-call+ mod, added the event (defttag :xdoc) to
xdoc/save-fancy.lisp.

Updated some documentation as per feedback from Warren Hunt:

- Updated :doc for encapsulate and progn to clarify that by default,
  each sub-event is printed.

- Updated :doc congruence so to clarify that initial example requires
  that set-equal be a known equivalence relation.

Updated :doc deftheory to note that deftheory events are never
redundant (thanks to David Rager for feedback leading to this change).

In :doc bookdata, mentioned community-books directory
books/tools/book-conflicts/.

Added to the item in :doc note-6-5 about world rollback for
certify-book, after confirming with Sol Swords, that he has reported a
nearly 20% reduction in time for certifying a certain large collection
of books.

Based on a chat with David Rager, clarified in :doc brr@ the
system-level nature of this utility and added a pointer to a defrec
for :ancestors.



git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2622 303c5b2c-9740-0410-b014-fd8f9db7be32