Revamp VL's handling of edge-triggered always blocks.
This replaces VL's previous "flopcode" stuff with some more sophisticated
transforms that can support a richer set of always blocks.  It should be mostly
transparent to users who are calling the always-top transform to deal with
always blocks (e.g., for users of vl-simplify).

One enhancement is that, via a new "edgesplit" transform, edge-triggered blocks
such as:

    always @(posedge clk)
    begin
       q1 <= d1;
       q2 <= d2;
    end

Are now supported.  Essentially such a block is split up into several blocks,
one for each register, which can then be synthesized separately.

The other major extension is that VL now supports certain always blocks with
multiple edge triggers.  The class of blocks supported are roughly those that
can be syntactically reduced to something like:

    always @(posedge clk1 or negedge clk2 or posedge clk3 or ...)
       if (clk1)
          q <= d1;
       else if (!clk2)
          q <= d2;
       else if (clk3)
          q <= d3;
       else ...

In other words, we support registers that use some priority encoding of the
clocks to choose between data inputs.  This class of registers includes, for
instance, registers with asynchronous resets such as:

    always @(posedge clk or posedge reset)
       if (reset)
           q <= 0;
       else
           q <= d;

And also fancier registers with, e.g., asynchronous set and reset signals, and
so forth.  In support of this change, ESIM now uses different register
primitives.

There are new VL systests to test out this functionality.  Also improved some
existing systests after discovering a problem with how randomization was being
implemented in them.

git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2659 303c5b2c-9740-0410-b014-fd8f9db7be32