Updating boothmul to use the newer version of stv-decomp-proofs.
Also adding an example use of specific inputs and hyps to show how
that approach can break.  The tutorial may not be the best place for
that, but I didn't want to create an additional
boothmul-noautoins.lisp

The hint required to make stv-decomp-theory work is now smaller, which
is kind of nice.

git-svn-id: http://acl2-books.googlecode.com/svn/trunk@2713 303c5b2c-9740-0410-b014-fd8f9db7be32